+++ dashboard/routes/case.js
@@ -44,41 +44,6 @@ router.get("/:cid", authenticateJWT, function(req, res, next) {
     })();
 });
 
-const checkAndUpdateBuildResult = async cid => {
-    const allCases = await caseService.getAllCasesByCid(cid);
-
-    let [ passedCount, failedCount, undeterminedCount, passedByIgnoringRectanglesCount ] = [ 0, 0, 0, 0 ];
-    for (const testCase of allCases) {
-        switch (testCase.caseResult) {
-            case "undetermined":
-                undeterminedCount += 1;
-                break;
-            case "failed":
-                failedCount += 1;
-                break;
-            case "passed":
-                passedCount += 1;
-                break;
-        }
-
-        if (testCase.comprehensiveCaseResult === "passed") {
-            passedByIgnoringRectanglesCount += 1;
-        }
-    }
-
-    await buildService.updateTestCaseCount(allCases[0].pid, allCases[0].bid, {
-        passed: passedCount,
-        failed: failedCount,
-        undeterminedCount: undeterminedCount,
-        passedByIgnoringRectangles: passedByIgnoringRectanglesCount
-    });
-
-    const buildResult = undeterminedCount ? "undetermined"
-        : failedCount > passedByIgnoringRectanglesCount ? "failed" : "passed";
-
-    await buildService.updateBuildResult(allCases[0].bid, buildResult);
-};
-
 const cleanTestCaseComprehensiveCaseResult = async (cid) => {
     /**
      * testCase.comprehensiveCaseResult is generated at Engine side, and should only be used in Dashboard page to detect
@@ -96,7 +61,7 @@ router.post("/pass/:cid", authenticateJWT, function(req, res, next) {
             await caseService.passCase(req.params.cid);
 
             await cleanTestCaseComprehensiveCaseResult(req.params.cid);
-            await checkAndUpdateBuildResult(req.params.cid);
+            await caseService.checkAndUpdateBuildResult(req.params.cid);
 
             res.redirect(`/case/${req.params.cid}`);
             console.log(`set case passed, cid=${req.params.cid}`);
@@ -114,7 +79,7 @@ router.post("/fail/:cid", authenticateJWT, function(req, res, next) {
             await caseService.failCase(req.params.cid);
 
             await cleanTestCaseComprehensiveCaseResult(req.params.cid);
-            await checkAndUpdateBuildResult(req.params.cid);
+            await caseService.checkAndUpdateBuildResult(req.params.cid);
 
             res.redirect(`/case/${req.params.cid}`);
             console.log(`set case failed, cid=${req.params.cid}`);
